refactor: Move DepTable to be with Manifest
