+++ crates/cargo-add/src/cargo/ops/cargo_add/manifest.rs
@@ -11,6 +11,31 @@ use super::dependency::Dependency;
 
 const DEP_TABLES: &[&str] = &["dependencies", "dev-dependencies", "build-dependencies"];
 
+/// Dependency table to add dep to
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum DepTable {
+    /// Used for building final artifact
+    Normal,
+    /// Used for testing
+    Development,
+    /// Used for build.rs
+    Build,
+    /// Used for building final artifact only on specific target platforms
+    Target(String),
+}
+
+impl DepTable {
+    /// Keys to the table
+    pub fn to_table(&self) -> Vec<&str> {
+        match self {
+            Self::Normal => vec!["dependencies"],
+            Self::Development => vec!["dev-dependencies"],
+            Self::Build => vec!["build-dependencies"],
+            Self::Target(target) => vec!["target", target, "dependencies"],
+        }
+    }
+}
+
 /// A Cargo manifest
 #[derive(Debug, Clone)]
 pub struct Manifest {
